+++ node/mdSchema.ts
@@ -12,8 +12,9 @@ export const ORGANIZATION_REQUEST_FIELDS = [
   'priceTables',
   'paymentTerms',
   'salesChannel',
+  'sellers',
 ]
-export const ORGANIZATION_REQUEST_SCHEMA_VERSION = 'v0.1.0'
+export const ORGANIZATION_REQUEST_SCHEMA_VERSION = 'v0.1.1'
 
 export const ORGANIZATION_DATA_ENTITY = 'organizations'
 export const ORGANIZATION_FIELDS = [
@@ -25,10 +26,11 @@ export const ORGANIZATION_FIELDS = [
   'priceTables',
   'salesChannel',
   'costCenters',
+  'sellers',
   'status',
   'created',
 ]
-export const ORGANIZATION_SCHEMA_VERSION = 'v0.0.7'
+export const ORGANIZATION_SCHEMA_VERSION = 'v0.0.8'
 
 export const COST_CENTER_DATA_ENTITY = 'cost_centers'
 export const COST_CENTER_FIELDS = [
@@ -139,6 +141,10 @@ export const schemas = [
           // deprecated
           type: 'array',
           title: 'Cost Centers',
+        },
+        sellers: {
+          type: 'array',
+          title: 'Sellers',
         },
         status: {
           type: 'string',