+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- added sellers to orgs
+- added ui modal options to settings
+- added sellers to organization request
+
+
 ## [0.27.0] - 2023-02-08
 
 ### Added